# fix(x/group/internal/orm): use proper map type to string not interface{}