--- a/x/group/internal/orm/indexer.go
+++ b/x/group/internal/orm/indexer.go
@@ -176,7 +176,7 @@ func multiKeyAddFunc(store storetypes.KVStore, secondaryIndexKey interface{}, ro
 
 // difference returns the list of elements that are in a but not in b.
 func difference(a, b []interface{}) ([]interface{}, error) {
-	set := make(map[interface{}]struct{}, len(b))
+	set := make(map[string]struct{}, len(b))
 	for _, v := range b {
 		bt, err := keyPartBytes(v, true)
 		if err != nil {